requirements.txt
Flask>=0.10.1
gunicorn>=19.3.0
PyICU>=1.9.2
latex>=0.6.1
